# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  client: "client",
  vendor: "vendor",
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  role: UserProfileRole;
  companyId?: number | null;
  companyName?: string | null;
}

export interface LoginResponse {
  token: string;
  user: UserProfile;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  client: "client",
  vendor: "vendor",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  companyId?: number | null;
  companyName?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  client: "client",
  vendor: "vendor",
} as const;

export interface CreateUserRequest {
  email: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  companyId?: number | null;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  client: "client",
  vendor: "vendor",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  role?: UpdateUserRequestRole;
  companyId?: number | null;
  isActive?: boolean;
  password?: string;
}

export type CompanyType = (typeof CompanyType)[keyof typeof CompanyType];

export const CompanyType = {
  client: "client",
  vendor: "vendor",
} as const;

export interface Company {
  id: number;
  name: string;
  type: CompanyType;
  isActive: boolean;
  createdAt: string;
}

export type CreateCompanyRequestType =
  (typeof CreateCompanyRequestType)[keyof typeof CreateCompanyRequestType];

export const CreateCompanyRequestType = {
  client: "client",
  vendor: "vendor",
} as const;

export interface CreateCompanyRequest {
  name: string;
  type: CreateCompanyRequestType;
}

export interface UpdateCompanyRequest {
  name?: string;
  isActive?: boolean;
}

export type JobRoleStatus = (typeof JobRoleStatus)[keyof typeof JobRoleStatus];

export const JobRoleStatus = {
  draft: "draft",
  pending_approval: "pending_approval",
  published: "published",
  closed: "closed",
} as const;

export interface JobRole {
  id: number;
  title: string;
  description?: string | null;
  skills?: string | null;
  status: JobRoleStatus;
  companyId: number;
  companyName: string;
  candidateCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRoleRequest {
  title: string;
  description?: string;
  skills?: string;
}

export interface UpdateRoleRequest {
  title?: string;
  description?: string;
  skills?: string;
}

export type UpdateRoleStatusRequestStatus =
  (typeof UpdateRoleStatusRequestStatus)[keyof typeof UpdateRoleStatusRequestStatus];

export const UpdateRoleStatusRequestStatus = {
  draft: "draft",
  pending_approval: "pending_approval",
  published: "published",
  closed: "closed",
} as const;

export interface UpdateRoleStatusRequest {
  status: UpdateRoleStatusRequestStatus;
}

export type CandidateStatus =
  (typeof CandidateStatus)[keyof typeof CandidateStatus];

export const CandidateStatus = {
  submitted: "submitted",
  screening: "screening",
  interview: "interview",
  offer: "offer",
  hired: "hired",
  rejected: "rejected",
} as const;

export interface Candidate {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  expectedSalary?: number | null;
  status: CandidateStatus;
  roleId: number;
  roleTitle: string;
  vendorCompanyId: number;
  vendorCompanyName: string;
  submittedAt: string;
  updatedAt: string;
}

export interface SubmitCandidateRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  expectedSalary?: number;
  roleId: number;
}

export type UpdateCandidateStatusRequestStatus =
  (typeof UpdateCandidateStatusRequestStatus)[keyof typeof UpdateCandidateStatusRequestStatus];

export const UpdateCandidateStatusRequestStatus = {
  submitted: "submitted",
  screening: "screening",
  interview: "interview",
  offer: "offer",
  hired: "hired",
  rejected: "rejected",
} as const;

export interface UpdateCandidateStatusRequest {
  status: UpdateCandidateStatusRequestStatus;
}

export interface Contract {
  id: number;
  candidateId: number;
  candidateName: string;
  roleTitle: string;
  vendorCompanyName: string;
  startDate: string;
  endDate?: string | null;
  dailyRate: number;
  isActive: boolean;
  createdAt: string;
}

export interface CreateContractRequest {
  candidateId: number;
  startDate: string;
  endDate?: string;
  dailyRate: number;
}

export interface Timesheet {
  id: number;
  contractId: number;
  candidateName: string;
  roleTitle: string;
  month: number;
  year: number;
  totalDays: number;
  totalAmount: number;
  submittedAt: string;
}

export interface SubmitTimesheetRequest {
  contractId: number;
  month: number;
  year: number;
  totalDays: number;
}

export type ListCandidatesParams = {
  roleId?: number;
};