# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ATS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyRequest"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: roleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: submitCandidate
      tags: [candidates]
      summary: Submit a candidate (Vendor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCandidateRequest"
      responses:
        "201":
          description: Candidate submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "409":
          description: Duplicate candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/{id}/status:
    patch:
      operationId: updateCandidateStatus
      tags: [candidates]
      summary: Update candidate status (Client HR)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateStatusRequest"
      responses:
        "200":
          description: Candidate status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, client, vendor]
        companyId:
          type: integer
          nullable: true
        companyName:
          type: string
          nullable: true
      required:
        - id
        - email
        - name
        - role
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [client, vendor]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - isActive
        - createdAt
    CreateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [client, vendor]
      required:
        - name
        - type
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        expectedSalary:
          type: number
          nullable: true
        status:
          type: string
          enum: [submitted, screening, interview, offer, hired, rejected]
        roleId:
          type: integer
        roleTitle:
          type: string
        vendorCompanyId:
          type: integer
        vendorCompanyName:
          type: string
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - status
        - roleId
        - roleTitle
        - vendorCompanyId
        - vendorCompanyName
        - submittedAt
        - updatedAt
    SubmitCandidateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        expectedSalary:
          type: number
        roleId:
          type: integer
      required:
        - firstName
        - lastName
        - email
        - roleId
    UpdateCandidateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, screening, interview, offer, hired, rejected]
      required:
        - status
    Contract:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: string
        roleTitle:
          type: string
        vendorCompanyName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        dailyRate:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - candidateName
        - roleTitle
        - vendorCompanyName
        - startDate
        - dailyRate
        - isActive
        - createdAt
    CreateContractRequest:
      type: object
      properties:
        candidateId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dailyRate:
          type: number
      required:
        - candidateId
        - startDate
        - dailyRate
...