# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateStatus } from "./candidateStatus";

export interface Candidate {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  expectedSalary?: number | null;
  status: CandidateStatus;
  roleId: number;
  roleTitle: string;
  vendorCompanyId: number;
  vendorCompanyName: string;
  submittedAt: Date;
  updatedAt: Date;
}