# lib/api-zod/src/generated/types/submitCandidateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitCandidateRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  expectedSalary?: number;
  roleId: number;
}